pyproject.toml
[tool.poetry]
name = "beer_finder"
version = "0.3.5.3"
description = "The app will help you find the best beer for you and your friends."
authors = ["Stas Cartoon <bycs@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"

Django = "4.2b1"
djangorestframework = "3.14.0"
drf-spectacular = "0.26.1"
psycopg = {extras = ["binary"], version = "3.1.8"}
gunicorn = "20.1.0"

python-telegram-bot = { extras = ["socks"], version = "20.1" }
pymongo = "4.3.3"
aiogram = "2.25.1"

requests = "2.28.2"
beautifulsoup4 = "4.11.2"
selenium = "4.8.2"
webdriver-manager = "3.8.5"

python-dotenv = "1.0.0"

[tool.poetry.group.dev.dependencies]
# code linters
black = "23.1.0"
isort = "5.12.0"

flake8 = "5.0.4"
flake8-black = "0.3.6"
flake8-builtins = "2.1.0"
flake8-django = "1.1.5"
flake8-isort = "6.0.0"
dotenv-linter = "0.4.0"

mypy = "1.1.1"
django-stubs = "1.16.0"
djangorestframework-stubs = "1.10.0"

pre-commit = "3.2.0"

# testing code
pytest = "7.2.2"
pytest-django = "4.5.2"
pytest-drf = "1.1.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
exclude = """
    ('
    |.idea/
    |.venv/
    |.vscode/
    ')
"""
